requirements.txt
fastapi
uvicorn
motor
pydantic
langchain
langchain-google-genai
langchain-huggingface
langchain-chroma
langchain-groq
chromadb
sentence-transformers
PyPDF2
pymongo
python-dotenv
